requirements/local.txt
-r base.txt

coverage==4.4.1
django-debug-toolbar==1.8
factory-boy==2.9.2
pytest-django==3.1.2